chore: Move Go signatures to top level signatures.
