+++ signatures/golang/examples/example.go
