+++ signatures/golang/export.go
