+++ signatures/golang/kubernetes_api_connection.go
@@ -4,8 +4,9 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/aquasecurity/tracee/signatures/helpers"
+
 	tracee "github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/helpers"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 )
