+++ signatures/golang/kubernetes_api_connection_test.go
@@ -3,8 +3,9 @@ package main
 import (
 	"testing"
 
+	"github.com/aquasecurity/tracee/signatures/signaturestest"
+
 	tracee "github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/signaturestest"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
