+++ signatures/golang/stdio_over_socket.go
@@ -3,8 +3,9 @@ package main
 import (
 	"fmt"
 
+	"github.com/aquasecurity/tracee/signatures/helpers"
+
 	tracee "github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/helpers"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 )
