+++ tracee-rules/benchmark/signature/golang/anti_debugging.go
@@ -3,8 +3,9 @@ package golang
 import (
 	"fmt"
 
+	"github.com/aquasecurity/tracee/signatures/helpers"
+
 	tracee "github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/helpers"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 )
